third_party/blink/web_tests/virtual/customizable-select-disabled/html5lib/generated/run-webkit01-write-expected.txt

CONSOLE MESSAGE: PASS
CONSOLE MESSAGE: FOO<span>BAR</span>BAZ
CONSOLE ERROR: A <select> tag was parsed within another <select> tag and was converted into </select>. This behavior will change in a future browser version. Please add the missing </select> end tag.
CONSOLE ERROR: A <select> tag was parsed within another <select> tag and was converted into </select>. This behavior will change in a future browser version. Please add the missing </select> end tag.
CONSOLE ERROR: A <select> tag was parsed within another <select> tag and was converted into </select>. This behavior will change in a future browser version. Please add the missing </select> end tag.
CONSOLE ERROR: A <select> tag was parsed within another <select> tag and was converted into </select>. This behavior will change in a future browser version. Please add the missing </select> end tag.
../resources/webkit01.dat: PASS
